[Fix] Stream Export: Fix bogus truncation of pattern cue points (no actual bug due to WAV being limited to 32bit anyway).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18303 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Mod2wave.cpp
@@ -1384,7 +1384,7 @@ void CDoWaveConvert::Run()
 			cues.reserve(patternCuePoints.size());
 			for(const auto &cue : patternCuePoints)
 			{
-				cues.push_back(static_cast<uint32>(cue.offset));
+				cues.push_back(cue.offset);
 			}
 			fileEnc->WriteCues(cues);
 		}
